+++ django_echarts/views/__init__.py
@@ -1,5 +1,4 @@
 # coding=utf8
-from __future__ import unicode_literals
 
 from .backend import EChartsBackendView
 from .base import EChartsMixin