+++ django_echarts/views/base.py
@@ -1,11 +1,11 @@
 # coding=utf8
 
-from __future__ import unicode_literals
 
 
 class EChartsMixin(object):
     """
     Base Mixin for build echarts instance
     """
+
     def get_echarts_instance(self, *args, **kwargs):
         pass