+++ example/demo/backend_views.py
@@ -1,6 +1,5 @@
 # coding=utf8
 
-from __future__ import unicode_literals
 
 from django.db.models import Count
 from pyecharts import Line, Pie, Page, Bar